+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/ControllerScanner.java
@@ -16,6 +16,7 @@ public class ControllerScanner {
 
     public Map<Class<?>, Object> getControllers() {
         Set<Class<?>> controllerClasses = reflections.getTypesAnnotatedWith(Controller.class);
+        System.out.println("size: "+controllerClasses.size());
         return instantiateController(controllerClasses);
     }
